+++ packages/realm-web/CHANGELOG.md
@@ -1,6 +1,18 @@
 ?.?.? Release notes (2020-??-??)
 =============================================================
 
+### Enhancements
+* None
+
+### Fixed
+* Fixed error `"function not found: 'argsTransformation'"` when calling `user.functions.callFunction('functionName', args)` [#3134](https://github.com/realm/realm-js/pull/3134)
+
+### Internal
+* None
+
+0.8.0 Release notes (2020-07-31)
+=============================================================
+
 ### Breaking changes
 * Renamed `Realm.app` to `Realm.getApp` to make it less conflicting with a local `app` variable.